Fixed bug 1 and 9
fixed negative in the range problem. Fixed the arrow keys not moving
cursor in the range textfields. Also fixed the button zooming. Now when
ever you don't have a cursor you can move the graph with the arrow keys
and zoom in an out via plus and minus keys. Works for the numberpad plus
and minus aswell. And press d anywhere to draw.